feat: Add resource delete.
